wip: manage how we show/hide the windows on each platform
