+++ crates/tauri/src/platform/mod.rs
@@ -1,3 +1,7 @@
 /// Platform specific implementation of things
+#[cfg(target_os = "linux")]
+pub mod linux;
 #[cfg(target_os = "macos")]
 pub mod mac;
+#[cfg(target_os = "windows")]
+pub mod windows;
